+++ CopilotKit/packages/runtime-client-gql/src/client/conversion.ts
@@ -204,8 +204,11 @@ export function loadMessagesFromJsonRepresentation(json: any[]): Message[] {
 
 function getPartialArguments(args: string[]) {
   try {
+    if (!args.length) return {};
+
     return JSON.parse(untruncateJson(args.join("")));
   } catch (e) {
+    console.error(e);
     return {};
   }
 }