+++ CopilotKit/packages/runtime/src/lib/runtime/remote-action-constructors.ts
@@ -47,6 +47,7 @@ export function constructLGCRemoteAction({
       threadId,
       nodeName,
       additionalMessages = [],
+      metaEvents,
     }: RemoteAgentHandlerParams): Promise<Observable<RuntimeEvent>> => {
       logger.debug({ actionName: agent.name }, "Executing LangGraph Platform agent");
 
@@ -67,7 +68,7 @@ export function constructLGCRemoteAction({
 
       try {
         const response = await execute({
-          logger,
+          logger: logger.child({ component: "remote-actions.remote-lg-action.streamEvents" }),
           deploymentUrl: endpoint.deploymentUrl,
           langsmithApiKey: endpoint.langsmithApiKey,
           agent,
@@ -81,6 +82,7 @@ export function constructLGCRemoteAction({
             description: action.description,
             parameters: JSON.parse(action.jsonSchema) as string,
           })),
+          metaEvents,
         });
 
         const eventSource = new RemoteLangGraphEventSource();
@@ -184,6 +186,7 @@ export function constructRemoteActions({
           threadId,
           nodeName,
           additionalMessages = [],
+          metaEvents,
         }: RemoteAgentHandlerParams): Promise<Observable<RuntimeEvent>> => {
           logger.debug({ actionName: agent.name }, "Executing remote agent");
 
@@ -219,6 +222,7 @@ export function constructRemoteActions({
                   description: action.description,
                   parameters: JSON.parse(action.jsonSchema),
                 })),
+                metaEvents,
               }),
             });