+++ CopilotKit/packages/shared/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from "./conditions";
 export * from "./errors";
 export * from "./json-schema";
 export * from "./random-id";