+++ examples/e2e/lib/config-helper.ts
@@ -1,5 +1,8 @@
 // lib/config-helper.ts
-import configs from "../app-configs.json";
+import configs from "../test-config.json";
+import dotenv from "dotenv";
+import path from "path";
+dotenv.config({ path: path.resolve(__dirname, "../.env") });
 
 // Project name constants with type safety
 export const PROJECT_NAMES = {
@@ -28,6 +31,7 @@ export interface ConfigMap {
 export interface TestVariant {
   name: string;
   queryParams: string;
+  isCloud?: boolean;
 }
 
 export type TestVariants = TestVariant[];
@@ -84,7 +88,7 @@ export const appendLGCVariants = (config: ConfigItem, variants: any[]) => {
     const newVariants = variants.map((variant) => {
       return {
         ...variant,
-        name: `${variant.name} (LGC Python)`,
+        name: `${variant.name} (LGC Python in-memory)`,
         queryParams: `${variant.queryParams}&lgcDeploymentUrl=${config.lgcPythonDeploymentUrl}`,
       };
     });
@@ -96,7 +100,7 @@ export const appendLGCVariants = (config: ConfigItem, variants: any[]) => {
     const newVariants = variants.map((variant) => {
       return {
         ...variant,
-        name: `${variant.name} (LGC JS \`main\`)`,
+        name: `${variant.name} (LGC JS in-memory)`,
         queryParams: `${variant.queryParams}&lgcDeploymentUrl=${config.lgcJSDeploymentUrl}`,
       };
     });
@@ -106,3 +110,31 @@ export const appendLGCVariants = (config: ConfigItem, variants: any[]) => {
 
   return appendedVariants;
 };
+
+export function getCopilotCloudVariants() {
+  const variants: TestVariants = [];
+
+  if (
+    process.env.COPILOT_CLOUD_PROD_RUNTIME_URL &&
+    process.env.COPILOT_CLOUD_PROD_PUBLIC_API_KEY
+  ) {
+    variants.push({
+      name: "Copilot Cloud (Production)",
+      queryParams: `?runtimeUrl=${process.env.COPILOT_CLOUD_PROD_RUNTIME_URL}&publicApiKey=${process.env.COPILOT_CLOUD_PROD_PUBLIC_API_KEY}`,
+      isCloud: true,
+    });
+  }
+
+  if (
+    process.env.COPILOT_CLOUD_STAGING_RUNTIME_URL &&
+    process.env.COPILOT_CLOUD_STAGING_PUBLIC_API_KEY
+  ) {
+    variants.push({
+      name: "Copilot Cloud (Staging)",
+      queryParams: `?runtimeUrl=${process.env.COPILOT_CLOUD_STAGING_RUNTIME_URL}&publicApiKey=${process.env.COPILOT_CLOUD_STAGING_PUBLIC_API_KEY}`,
+      isCloud: true,
+    });
+  }
+
+  return variants;
+}