+++ examples/e2e/tests/coagents-qa-native-demo.spec.ts
@@ -7,12 +7,13 @@ import {
   PROJECT_NAMES,
   TestVariants,
   appendLGCVariants,
+  getCopilotCloudVariants,
 } from "../lib/config-helper";
 
 const variants: TestVariants = [
   { name: "OpenAI", queryParams: "?coAgentsModel=openai" },
   { name: "Anthropic", queryParams: "?coAgentsModel=anthropic" },
-  // { name: "Google Generative AI", queryParams: "?coAgentsModel=google_genai" }, // seems broken
+  ...getCopilotCloudVariants(),
 ];
 
 const allConfigs = getConfigs();
@@ -22,18 +23,25 @@ const qaConfigs = filterConfigsByProject(
 );
 const groupedConfigs = groupConfigsByDescription(qaConfigs);
 
+export const cloudVariants = variants.filter((variant) => variant.isCloud);
+export const nonCloudVariants = variants.filter((variant) => !variant.isCloud);
+
+test.describe.configure({ mode: 'parallel' });
+
 Object.entries(groupedConfigs).forEach(([projectName, descriptions]) => {
   test.describe(`${projectName}`, () => {
     Object.entries(descriptions).forEach(([description, configs]) => {
       test.describe(`${description}`, () => {
         configs.forEach((config) => {
-          appendLGCVariants(
-            {
-              ...config,
-              lgcJSDeploymentUrl: config.lgcJSDeploymentUrl,
-            },
-            variants
-          ).forEach((variant) => {
+          [
+            ...appendLGCVariants(
+              {
+                ...config,
+              },
+              nonCloudVariants
+            ),
+            ...cloudVariants,
+          ].forEach((variant) => {
             test(`Test ${config.description} with variant ${variant.name}`, async ({
               page,
             }) => {
@@ -56,6 +64,12 @@ Object.entries(groupedConfigs).forEach(([projectName, descriptions]) => {
                 page,
                 "write an email to the CEO of OpenAI asking for a meeting"
               );
+
+              // First interaction will bring up interrupt interface
+              await page
+                  .getByPlaceholder("Your name")
+                  .fill('CopilotKit Automation');
+              await page.locator('button:has-text("Submit")').click();
 
               const cancelMessage = page.locator(
                 '[data-test-id="email-cancel-message"]'