+++ CopilotKit/packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @copilotkit/sdk-js
 
+## 1.5.15-next.4
+
+### Patch Changes
+
+- Updated dependencies [7b3141d]
+  - @copilotkit/shared@1.5.15-next.4
+
+## 1.5.15-next.3
+
+### Patch Changes
+
+- @copilotkit/shared@1.5.15-next.3
+
+## 1.5.15-next.2
+
+### Patch Changes
+
+- @copilotkit/shared@1.5.15-next.2
+
+## 1.5.15-next.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.5.15-next.1
+
+## 1.5.15-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.5.15-next.0
+
+## 1.5.14
+
+### Patch Changes
+
+- Updated dependencies [0061f65]
+  - @copilotkit/shared@1.5.14
+
 ## 1.5.14-next.0
 
 ### Patch Changes