+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @copilotkit/shared
 
+## 1.5.15-next.4
+
+### Patch Changes
+
+- 7b3141d: - feat(interrupt): support LG interrupt with useLangGraphInterrupt hook
+  - chore(interrupt): add e2e test to interrupt functionality
+  - feat(interrupt): add support for multiple interrupts and conditions
+
+## 1.5.15-next.3
+
+## 1.5.15-next.2
+
+## 1.5.15-next.1
+
+## 1.5.15-next.0
+
+## 1.5.14
+
+### Patch Changes
+
+- 0061f65: - feat: allows dev mode for cloud onboarding flow
+
 ## 1.5.14-next.0
 
 ### Patch Changes